refactor: move canvas and loader services into visualization package
