+++ visualization/infrastructure/qgis_layer_gateway.py
@@ -4,10 +4,10 @@ from qgis.core import QgsProject
 
 logger = logging.getLogger(__name__)
 
-from ...map_canvas_service import MapCanvasService
 from ...mapbox_config import TILE_MODE_RASTER
-from ...project_layer_loader import ProjectLayerLoader
 from .background_map_service import BackgroundMapService
+from .map_canvas_service import MapCanvasService
+from .project_layer_loader import ProjectLayerLoader
 from .temporal_service import TemporalService
